+++ packages/Table/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - default border has been changed from grid to horizontal. [@kaan.darcey](https://github.com/KDarcey).
 
+### Added
+
+- BREAKING: Propagate `<Table.ColumnDefinition>` `moreProps` to the `<th>` elements. This could introduce a breaking change since props that were applied to `<td>` elements will now ALSO be applied to `<th>` elements. [@mikrotron](https://github.com/mikrotron)
+
 ## [0.1.3] - 2020-08-27
 
 - Add default font styles on th [@tristanjasper](https://github.com/tristanjasper).